refactor(Refactor main `run_game()` in wizard_main.py): Refactored the `run_game` function into two seperate helper functions
Placed the loop to handle events into it's own helper function as later on it'd be easiler to expand with different key events. Then I placed the code for updating the screen into it's own function. This all made the body of the main loop  more simplier to read and understand exactly what it needs to do now.