build(Spell class and facing direction): Build spell class for fireball in new file, and added new feature to make the direction going also make the face of wizard face that way.
I came to a thought with this project, that in the original Space Invader game bullets only need to shoot upwards. In my current verison with 4 directions of travel, I needed a way to allow the  spells to know which way the wizard is facing and keep moving in 'that' direction.  In wizard.py, I gave the wizard a set of known direction values. Then in `wizard.update()`, those values are updated using the coordinate system of Pygame whenever the wizard, it remembers the last direction it moved. This way the spell fired can then use that stored direction when it is called.