feat(Random starting enemy): Add randomness to the game by making the enemies spawn along certain sides.
I updated the monster’s init so that each one now spawns at a random position along the top, left, or right edges of the screen. I did this by selecting a random edge and then assigning the monster’s rect coordinates based on that choice. This makes the game less predictable and prevents enemies from always appearing in the same location, this should make the game more challenging.