feat(Game Stats, and Handle Monster Collisions): Add statistic feature to end the game later on with lives, add time module, and function to reset the game after a hit.
Created a statistics file to keep track of times a wizard is hit. Then in the main file, the time module was added to pause the gameplay for a bit when a hit occurs.  When a hit occurs then, it'll remake the monsters and clear the screen from any spells to give it a fresh start.