feat(Game Over): Add in main functions to send a reset message, and reset game. In wizard.py a function  added to center wizard.
When the game ends, right away a message will show on screen to press the SPACE to reset the game.  I placed the reset button to the same one to shot the spells, as it made sense since your thumb was already on it.  A simple if loop in _check_keydown() was made for that to call the reset game function. One extra funciton that wasn't strictly needed was the center_wizard_position() but it made the game a bit more fair. After reseting, if you were too close to the edge, you could loose a life too quickly.